Address code-review findings: FFI soundness, Windows races, cursor nonce, CI
Correctness / soundness:
- Convert the cursor FFI layer to raw-pointer reads/writes: no &/&mut is
  formed over caller memory (uninitialized fh_cursor storage, possibly
  aliased ip/sp output pointers), and fh_capture_context now writes all
  five context words so no uninitialized word sits behind a reference.
- Give each slot claim a nonce, verified on step/get_reg and retired by
  CAS at release: a stale cursor copy or a double-fini racing the slot's
  next owner is now a guaranteed no-op instead of freeing another walk's
  slot (hazard clear + in_use corruption). Regression test added.
- Windows: pin each module (GetModuleHandleExW FROM_ADDRESS + FreeLibrary)
  while parsing/copying its PE image, so a concurrent FreeLibrary cannot
  unmap it under us; grow the EnumProcessModules buffer in a retry loop
  and keep the previous snapshot on enumeration failure instead of
  treating an empty result as "everything unloaded"; fold the PE header
  TimeDateStamp into the module fingerprint (size alone cannot tell two
  DLLs apart across base-address reuse).
- Count cache-sweep progress in module *mutations* (via a WriterUnw
  counting facade) rather than publishes: one publish can carry many
  mutations, so the u16 modules-generation wrap bound now actually holds.

Packaging / docs:
- rust-version 1.88 (naked_asm requires it; LazyLock needs 1.80 — the
  declared 1.78 could never build).
- Add the missing fh_jit_module_count / fh_jit_register_failures stubs
  for unsupported arches so consumers always link.
- Allow clippy::mut_from_ref on Slot::inner_mut with justification;
  fix all remaining clippy warnings (tree is now clippy-clean).
- Document the longjmp-over-Rust-frames caveat (POF rule): exact stack
  bounds are the supported configuration, longjmp recovery is best-effort.
- README: the Julia integration lives on the JuliaLang branch
  gb/framehop-unwinder (the in-repo patch was dropped in 2dacae3);
  add fh_cursor_init_bounds to the API mapping table.
- Add CI (build/test/clippy on x86_64+arm64 Linux and macOS, TSan run of
  the concurrency stress test, MSRV check, cross-checks for all six
  targets) and tests/run_c_smoke.sh to exercise the real C ABI.

+++ src/modules/elf.rs
@@ -85,7 +85,7 @@ extern "C" fn callback(
     let info = unsafe { &*info };
     let collector = unsafe { &mut *(data as *mut Collector) };
 
-    let bias = info.dlpi_addr as u64;
+    let bias = info.dlpi_addr;
     if info.dlpi_phdr.is_null() || info.dlpi_phnum == 0 {
         return 0;
     }
@@ -100,8 +100,8 @@ extern "C" fn callback(
     for ph in phdrs {
         match ph.p_type {
             PT_LOAD => {
-                let lo = bias + ph.p_vaddr as u64;
-                let hi = lo + ph.p_memsz as u64;
+                let lo = bias + ph.p_vaddr;
+                let hi = lo + ph.p_memsz;
                 if lo < avma_lo {
                     avma_lo = lo;
                 }
@@ -110,8 +110,8 @@ extern "C" fn callback(
                 }
                 if (ph.p_flags & PF_X) != 0 && text_svma.is_none() {
                     text_svma = Some(Range {
-                        start: ph.p_vaddr as u64,
-                        end: ph.p_vaddr as u64 + ph.p_memsz as u64,
+                        start: ph.p_vaddr,
+                        end: ph.p_vaddr + ph.p_memsz,
                     });
                 }
             }
@@ -135,7 +135,7 @@ extern "C" fn callback(
         unsafe { CStr::from_ptr(info.dlpi_name) }.to_bytes()
     };
     let (hdr_vaddr, hdr_memsz) = match hdr_phdr {
-        Some(p) => (p.p_vaddr as u64, p.p_memsz as u64),
+        Some(p) => (p.p_vaddr, p.p_memsz),
         None => (0, 0),
     };
     let fp = super::fingerprint_of(&[hdr_vaddr, hdr_memsz, avma_hi - avma_lo], name_bytes);
@@ -152,7 +152,7 @@ extern "C" fn callback(
         None => return 0, // no unwind info; framehop will fp-fallback for this module
     };
 
-    let hdr_avma = bias + hdr_phdr.p_vaddr as u64;
+    let hdr_avma = bias + hdr_phdr.p_vaddr;
     let hdr_len = hdr_phdr.p_memsz as usize;
     if hdr_len < 4 {
         return 0;
@@ -221,8 +221,8 @@ fn containing_segment_end(phdrs: &[libc::Elf64_Phdr], bias: u64, addr: u64) -> O
         if ph.p_type != PT_LOAD {
             continue;
         }
-        let lo = bias + ph.p_vaddr as u64;
-        let hi = lo + ph.p_memsz as u64;
+        let lo = bias + ph.p_vaddr;
+        let hi = lo + ph.p_memsz;
         if addr >= lo && addr < hi {
             return Some(hi);
         }
